requirements.txt
streamlit>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
ollama>=0.1.7
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
pygments>=2.16.0

# Optional for enhanced language detection (has TensorFlow dependencies)
# guesslang>=2.2.1